% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{e}
\alias{e}
\title{Environment for useful variables.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
e
}
\description{
Environment for useful variables.
}
\keyword{datasets}
