% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main analysis steering function}
\usage{
main(
  expt = "default",
  ice_sources = c("GrIS", "AIS", "Glaciers"),
  years = 2100,
  dataset = "main",
  N_temp = 1000L,
  N_FACTS = 2237L,
  outdir = "results",
  temp_prior = "FAIR",
  fair_ssps = NA,
  mean_temp = FALSE,
  gamma0_prior = "joint",
  mean_melt = FALSE,
  collapse_prior = "both",
  risk_averse = FALSE,
  min_res_is = c(NA, NA),
  select_ism = "all",
  select_gcm = "all",
  history_match = FALSE,
  exclude_open = FALSE,
  impute_high = FALSE,
  do_model_comp = FALSE,
  do_covar_fn = NA,
  do_covar_alpha = NA,
  packagename = "emulandice"
)
}
\arguments{
\item{expt}{Analysis to run: e.g. "SA", "timeseries"}

\item{ice_sources}{Ice sources: GrIS, AIS, Glaciers}

\item{years}{Year(s) to predict: default is 2100, time series is 2015:2100}

\item{dataset}{Forcing dataset: 2019, main, IPCC, FACTS}

\item{N_temp}{Number of climate values in prior: 501 code testing, 1000L default for tests, 5000L projections, over-ridden for timeseries}

\item{N_FACTS}{Number of  FAIR time series samples in forcing file passed by FACTS}

\item{outdir}{output directory}

\item{temp_prior}{Climate ensemble for prior: FAIR, CMIP6}

\item{fair_ssps}{Restrict FAIR SSPs run: NA; or e.g. c("SSP126", "SSP585"); must be set for IPCC timeseries runs}

\item{mean_temp}{Use mean temperature value or ice sheets: T/F}

\item{gamma0_prior}{Gamma0 prior distribution for AIS: "joint", "MeanAnt", "PIGL", "unif", "unif_high"}

\item{mean_melt}{Use mean kappa/gamma0 value for ice sheets: T/F}

\item{collapse_prior}{Ice shelf collapse prior: both, on, off}

\item{risk_averse}{Risk-averse settings: T/F}

\item{min_res_is}{Set minimum spatial resolution for ice sheets in km (GIS, AIS): default c(8, 32)}

\item{select_ism}{Select subset of ice sheet models by defined list: balanced, single, single_low, high_sensitivity, larmip}

\item{select_gcm}{Select subset of gcms by defined list: high_five, high_two}

\item{history_match}{Select subset of ice sheet models using IMBIE observations: T/F}

\item{exclude_open}{Exclude open parameterisation Antarctic ice sheet models: T/F}

\item{impute_high}{Use higher value of basal melt for open parameterisation Antarctic models: T/F}

\item{do_model_comp}{Run stepwise model comparison: T/F}

\item{do_covar_fn}{Set fixed covariance function for all regions: matern_5_2, matern_3_2, pow_exp}

\item{do_covar_alpha}{Set fixed pow_exp exponent for all regions: 0.1, 1.0, 1.9}

\item{packagename}{Set package name}
}
\description{
Main analysis steering function
}
