% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mme.R
\name{plot_mme}
\alias{plot_mme}
\title{Plots and outputs from raw simulation data}
\usage{
plot_mme(is, reg)
}
\arguments{
\item{is}{Ice source}

\item{reg}{Ice source region}
}
\description{
Plots and outputs from raw simulation data
}
