% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_forcing}
\alias{read_forcing}
\title{Read forcing data from CSV}
\usage{
read_forcing(
  scenario_list,
  temp_prior,
  N_temp,
  climate_prior_kde,
  mean_temp,
  old_data
)
}
\arguments{
\item{scenario_list}{Scenarios to read}

\item{temp_prior}{Which ensemble}

\item{N_temp}{Number of GSAT samples wanted}

\item{climate_prior_kde}{Whether to smooth prior before sampling}

\item{mean_temp}{Whether to use mean of GSAT prior instead of sampling}

\item{old_data}{Whether to read old forcing CSV file (2019) for backwards comparisons
Read in forcing CSV}
}
\description{
Read forcing data from CSV
}
