% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_sealevel}
\alias{read_sealevel}
\title{Read in sea level CSV}
\usage{
read_sealevel(min_res, old_data, sle.filename)
}
\arguments{
\item{min_res}{Arg: vector of minimum grid resolution in km for GrIS, AIS models}

\item{old_data}{Whether data file is the old one (2019) for backward comparions (N.B. the data is wrong!)}

\item{sle.filename}{Name of CSV to read in}
}
\description{
Read in sea level CSV
}
