% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Sensitivity analysis: dependence of results on temperature and ice sheet parameters}
\usage{
sensitivity_analysis(year, ice_source, region, sim, emulator, N_melt = 1)
}
\arguments{
\item{year}{Year}

\item{ice_source}{Ice source}

\item{region}{Ice source region}

\item{sim}{Simulation data for plotting}

\item{emulator}{Emulator object for predicting}

\item{N_melt}{Number of samples in plots sampling melt (not used in MAIN plots)}
}
\description{
Sensitivity analysis: dependence of results on temperature and ice sheet parameters
}
